RUMM-1541 Simplify repo structure by deleting example application and its test target
This new structure is created with `pod lib create` and selecting `No` for creating example "app project".
That produced a new sceleton which includes example "framework project". This shuold no longer
lead to misunderstanding the targets (now we have less targets) and there is no need to re-generate
any schema manually after `pod install` (as it used to be).